requirements.txt
aes-everywhere==1.2.10
alembic==1.8.0
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
bcrypt==3.2.2
Brotli==1.0.9
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.0
Flask-Compress==1.12
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-Pydantic==0.9.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.14.0
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
gunicorn==20.0.4
greenlet==1.1.2
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==4.6.1
Mako==1.2.1
MarkupSafe==1.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
numpy==1.21.6
packaging==21.3
pluggy==1.0.0
proto-plus==1.20.6
protobuf==3.20.1
psycopg2==2.9.3
py==1.11.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.39
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.10
Werkzeug==1.0.1
wrapt==1.14.1
zipp==3.8.0
